+++ CHANGELOG.md
@@ -8,8 +8,8 @@ Adding support for non-blocking futures and some minor API cleanup.
 
 ### New Features
 
-* Added support for non-blocking Future responses based on `guzzlehttp/ring`
-  and `guzzlehttp/ring-client`. You can still use the `sendAll()` and `batch()`
+* Added support for non-blocking Future responses based on
+  `guzzlehttp/guzzle-ring`. You can still use the `sendAll()` and `batch()`
   functions in exactly the same way, but you now also have access to the
   `future` request option which creates a future request if possible.
 * You can now create a default adapter based on the environment using a public
@@ -58,8 +58,8 @@ interfaces.
   `GuzzleHttp\Message\MessageInterface::getHeader`. If you want to get a header
   value as an array, then use the newly added ``getHeaderLines()`` method of
   ``MessageInterface``.
-* ``GuzzleHttp\Utils::batch`` now returns an ordred indexed array of responses
-  and exceptions rather than an ``SplObjectStorage`` object.
+* ``GuzzleHttp\Utils::batch`` now returns a `GuzzleHttp\BatchResults` object
+  instead of an SplObjectStorage.
 * ``GuzzleHttp\Message\MessageFactory`` no longer allows subclasses to add
   custom request options using double-dispatch. Instead, you should now provide
   an associative array to the constructor which is a mapping of the request
@@ -70,6 +70,15 @@ interfaces.
   This control mechanism was used to stop a transfer of parallel requests from
   completing. This can now be handled by throwing the exception or by
   cancelling a pool of requests or each outstanding future request individually.
+* Updated to "GuzzleHttp\Streams" 3.0.
+    * Removed functions.php from Streams
+    * `GuzzleHttp\Stream\StreamInterface::getContents()` no longer accepts a
+      `maxLen` parameter.
+    * ``GuzzleHttp\Stream\Stream::__construct``,
+      ``GuzzleHttp\Stream\Stream::factory``, and
+      ``GuzzleHttp\Stream\Utils::create`` no longer accept a size in the second
+      argument. They now accept an associative array of options, including the
+      "size" key and "metadata" key which can be used to provide custom metadata.
 
 4.2.2 (2014-09-08)
 ------------------